RevertBackToBuiltInConverter now works as expected
