+++ src/Tests/Framework/ManagedElasticsearch/SourceSerializers/TestJoinFieldJsonConverter.cs
@@ -1,6 +1,7 @@
 using System;
 using System.IO;
 using System.Linq;
+using System.Text;
 using Elasticsearch.Net;
 using Nest;
 using Newtonsoft.Json;
@@ -8,7 +9,6 @@ using Newtonsoft.Json.Linq;
 
 namespace Tests.Framework.ManagedElasticsearch.SourceSerializers
 {
-
 	internal class RevertBackToBuiltinSerializer : JsonConverter
 	{
 		private readonly IElasticsearchSerializer _builtInSerializer;
@@ -29,7 +29,8 @@ namespace Tests.Framework.ManagedElasticsearch.SourceSerializers
 
 		public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
 		{
-			using (var ms = new MemoryStream(reader.ReadAsBytes()))
+			var token = JToken.ReadFrom(reader);
+			using (var ms = new MemoryStream(Encoding.UTF8.GetBytes(token.ToString())))
 				return _builtInSerializer.Deserialize(objectType, ms);
 		}
 
@@ -39,66 +40,5 @@ namespace Tests.Framework.ManagedElasticsearch.SourceSerializers
 
 		public override bool CanConvert(Type objectType) => TypesThatCanAppearInSource.Contains(objectType);
 
-	}
-
-
-
-
-	/// <summary>
-	/// Hack to get our tests to serialize JoinFields correctly even when
-	/// using custom source serializer
-	/// </summary>
-	internal class TestJoinFieldJsonConverter :JsonConverter
-	{
-		public override bool CanRead => true;
-		public override bool CanWrite => true;
-
-		public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
-		{
-			if (reader.TokenType == JsonToken.String)
-			{
-				var parent = reader.Value.ToString();
-				return new JoinField(new JoinField.Parent(parent));
-			}
-			var jObject = JObject.Load(reader);
-			if (jObject.Properties().Any(p=>p.Name == "parent"))
-				using(var childReader =  jObject.CreateReader())
-					return (JoinField)serializer.Deserialize<JoinField.Child>(childReader);
-
-			using(var parentReader = jObject.CreateReader())
-				return (JoinField)serializer.Deserialize<JoinField.Parent>(parentReader);
-		}
-
-		public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
-		{
-			var join = value as JoinField;
-			if (join == null)
-			{
-				writer.WriteNull();
-				return;
-			}
-			join.Match(
-				p => serializer.Serialize(writer, GetName(p.Name)),
-				c =>
-				{
-					writer.WriteStartObject();
-					WriteProperty(writer, serializer, "name", GetName(c.Name));
-					WriteProperty(writer, serializer, "parent", GetId(c.Parent));
-					writer.WriteEndObject();
-				}
-			);
-		}
-
-		private string GetName(RelationName relation) => TestClient.DefaultInMemoryClient.Infer.RelationName(relation);
-		private string GetId(IUrlParameter id) => id.GetString(TestClient.DefaultInMemoryClient.ConnectionSettings);
-
-		public static void WriteProperty(JsonWriter writer, JsonSerializer serializer, string propertyName, object value)
-		{
-			if (value == null) return;
-			writer.WritePropertyName(propertyName);
-			serializer.Serialize(writer, value);
-		}
-
-		public override bool CanConvert(Type objectType) => typeof(JoinField) == objectType;
 	}
 }
